+++ packages/prismjs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @shallot-ui/prismjs
 
+## 3.0.3
+
+### Patch Changes
+
+- Adds color options to get input styles
+- Updated dependencies
+  - @shallot-ui/web@0.22.3
+
+## 3.0.2
+
+### Patch Changes
+
+- Fixes input sizing
+- Updated dependencies
+  - @shallot-ui/web@0.22.2
+
 ## 3.0.1
 
 ### Patch Changes