+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shallot-ui/theme
 
+## 0.22.3
+
+### Patch Changes
+
+- Adds color options to get input styles
+
+## 0.22.2
+
+### Patch Changes
+
+- Fixes input sizing
+
 ## 0.22.1
 
 ### Patch Changes