+++ packages/web/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shallot-ui/web
 
+## 0.22.3
+
+### Patch Changes
+
+- Adds color options to get input styles
+- Updated dependencies
+  - @shallot-ui/theme@0.22.3
+  - @shallot-ui/core@0.22.3
+
+## 0.22.2
+
+### Patch Changes
+
+- Fixes input sizing
+- Updated dependencies
+  - @shallot-ui/theme@0.22.2
+  - @shallot-ui/core@0.22.2
+
 ## 0.22.1
 
 ### Patch Changes